Add more JavaDoc header comments
