+++ data/star.txt
@@ -56,4 +56,5 @@ E | 0 | kabw | from Sep 30 2020 03:00 to Dec 4 2020 04:00
 E | 0 | kwhbvf | 13-09-2020
 E | 0 | lwkvjnk | whee
 E | 0 | ljvn | wurvbh
-T | 0 | run | #whee
+T | 0 | run
+T | 0 | run
